requirements.txt
Django==3.0.8
djangorestframework==3.11.0
psycopg2==2.8.5
requests==2.24.0
